// modules/accounts-management/modules/t365-movil/enum/url-collection.enum.ts
export enum AM365UrlCollection {
  HOME = 'accounts-management/365-movil',
  DETAIL = 'accounts-management/365-movil/detail',
  CREATE = 'accounts-management/365-movil/create',
  CREATE_CONFIRMATION = 'accounts-management/365-movil/create/confirmation',
  CREATE_VOUCHER = 'accounts-management/365-movil/create/voucher',
  DELETE = 'accounts-management/365-movil/delete',
  DELETE_CONFIRMATION = 'accounts-management/365-movil/delete/confirmation',
  DELETE_VOUCHER = 'accounts-management/365-movil/delete/voucher',
  UPDATE = 'accounts-management/365-movil/update',
  UPDATE_CONFIRMATION = 'accounts-management/365-movil/update/confirmation',
  UPDATE_VOUCHER = 'accounts-management/365-movil/update/voucher',
}