// modules/accounts-management/modules/t365-movil/enum/route-protected.enum.ts
export enum M365RouteProtectedParameters {
  DETAIL = 'm365:detail',
  CREATE = 'm365:create',
  CREATE_CONFIRM = 'm365:create_confirm',
  CREATE_VOUCHER = 'm365:create_voucher',
  DELETE = 'm365:delete',
  DELETE_CONFIRM = 'm365:delete_confirm',
  DELETE_VOUCHER = 'm365:delete_voucher',
  UPDATE = 'm365:update',
  UPDATE_CONFIRM = 'm365:update_confirm',
  UPDATE_VOUCHER = 'm365:update_voucher',
}